Run quotes thru smartypants
This proved more trouble than anticipated, dealing with """ and "" was
something of a pain. But also how to encode it, react is sensitive in
rendering html entities and altho there are methods to deal with it, in
the end i decided that the easiest for now is to simply replace with the
proper unicode character, so that is what i did.