Move handling of problemdetails to handler
Not sure if this is better, but, hey, you have to try things out, the
thinking here is that since this, as in the api, should be a super
simple passthru i want to simplify the boilerplate code as much as possible.
So handling everything but the happy path in places other than the
controller seems a better way to go.