Separate State and Api in the Provider
Having thought about this and read some articles around it, it made
sense to split state and api into seperated usings. This mostly to
avoid rerenders as much as possible. This is of course a massive overkill
for the purpose of this application and as an astute reader might notice
there is way more plumbing than actual code, but this is okay, this
project was always about testing new things and that sometimes mean
overengineering stuff.