Ensure that Cache-Control is set
So, having read up about what should and should not be cached in a spa
application with proper cachebusting this seems to be the wisdom of
the day, everything that is cache busted keep for a year and everything
else gets no-cache. It took a while to realize that with the fallback
to index.html, which is needed for anything to work properly, you need
to set not only the cache policy on the static files but also on the
fallback since those request doesnt go thru the static files. But you
learn something new every day.